# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Concrete business API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List all quotes
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a new quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteBody"
      responses:
        "201":
          description: Quote created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteWithLineItems"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{id}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get a single quote with its line items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote with line items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteWithLineItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a quote (status, line items, notes)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteWithLineItems"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{id}/status:
    patch:
      operationId: updateQuoteStatus
      tags: [quotes]
      summary: Update quote status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuoteStatusBody"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{id}/line-items:
    post:
      operationId: addLineItem
      tags: [quotes]
      summary: Add a line item to a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertLineItemBody"
      responses:
        "201":
          description: Line item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteLineItem"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{id}/line-items/{itemId}:
    put:
      operationId: updateLineItem
      tags: [quotes]
      summary: Update a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertLineItemBody"
      responses:
        "200":
          description: Updated line item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteLineItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteLineItem
      tags: [quotes]
      summary: Remove a line item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Quote:
      type: object
      properties:
        id:
          type: integer
        quoteRequestId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
        status:
          type: string
          enum: [draft, sent, accepted, declined]
        totalAmount:
          type: number
        expiryDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - customerName
        - customerEmail
        - status
        - totalAmount
        - createdAt
        - updatedAt
    QuoteLineItem:
      type: object
      properties:
        id:
          type: integer
        quoteId:
          type: integer
        description:
          type: string
        unit:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required:
        - id
        - quoteId
        - description
        - unit
        - quantity
        - unitPrice
        - total
    QuoteWithLineItems:
      allOf:
        - $ref: "#/components/schemas/Quote"
        - type: object
          properties:
            lineItems:
              type: array
              items:
                $ref: "#/components/schemas/QuoteLineItem"
          required:
            - lineItems
    CreateQuoteBody:
      type: object
      properties:
        quoteRequestId:
          type: integer
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
        expiryDate:
          type: string
        notes:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/UpsertLineItemBody"
      required:
        - customerName
        - customerEmail
    UpdateQuoteBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, accepted, declined]
        notes:
          type: string
          nullable: true
        expiryDate:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/UpsertLineItemBody"
    UpdateQuoteStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, accepted, declined]
      required:
        - status
    UpsertLineItemBody:
      type: object
      properties:
        description:
          type: string
        unit:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
      required:
        - description
        - unit
        - quantity
        - unitPrice
...