# lib/api-zod/src/generated/types/createQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Concrete business API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LineItemInput } from "./lineItemInput";

export interface CreateQuoteBody {
  quoteRequestId?: number | null;
  customerName: string;
  customerEmail: string;
  expiryDate?: string;
  notes?: string | null;
  lineItems?: LineItemInput[];
}