# lib/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Concrete business API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteStatus } from "./quoteStatus";

export interface Quote {
  id: number;
  quoteRequestId?: number | null;
  customerName: string;
  customerEmail: string;
  status: QuoteStatus;
  totalAmount: number;
  expiryDate?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}