# lib/api-zod/src/generated/types/updateQuoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Concrete business API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LineItemInput } from "./lineItemInput";
import type { UpdateQuoteBodyStatus } from "./updateQuoteBodyStatus";

export interface UpdateQuoteBody {
  status?: UpdateQuoteBodyStatus;
  notes?: string | null;
  lineItems?: LineItemInput[];
}